Sema: Shuffle some code between TypeCheckProtocol.cpp and TypeCheckProtocolInference.cpp
