+++ lib/Sema/TypeCheckProtocol.cpp
@@ -78,26 +78,6 @@ namespace {
   }
 }
 
-AssociatedTypeDecl *
-swift::getReferencedAssocTypeOfProtocol(Type type, ProtocolDecl *proto) {
-  if (auto dependentMember = type->getAs<DependentMemberType>()) {
-    if (auto assocType = dependentMember->getAssocType()) {
-      if (dependentMember->getBase()->isEqual(proto->getSelfInterfaceType())) {
-        // Exact match: this is our associated type.
-        if (assocType->getProtocol() == proto)
-          return assocType;
-
-        // Check whether there is an associated type of the same name in
-        // this protocol.
-        if (auto *found = proto->getAssociatedType(assocType->getName()))
-          return found;
-      }
-    }
-  }
-
-  return nullptr;
-}
-
 namespace {
   /// The kind of variance (none, covariance, contravariance) to apply
   /// when comparing types from a witness to types in the requirement
@@ -2880,59 +2860,6 @@ ConformanceChecker::ConformanceChecker(
 
 ConformanceChecker::~ConformanceChecker() {}
 
-
-TinyPtrVector<AssociatedTypeDecl *>
-ReferencedAssociatedTypesRequest::evaluate(Evaluator &eval,
-                                           ValueDecl *req) const {
-  // Collect the set of associated types rooted on Self in the
-  // signature. Note that for references to nested types, we only
-  // want to consider the outermost dependent member type.
-  //
-  // For example, a requirement typed '(Iterator.Element) -> ()'
-  // is not considered to reference the associated type 'Iterator'.
-  TinyPtrVector<AssociatedTypeDecl *> assocTypes;
-
-  class Walker : public TypeWalker {
-    ProtocolDecl *Proto;
-    llvm::TinyPtrVector<AssociatedTypeDecl *> &assocTypes;
-    llvm::SmallPtrSet<AssociatedTypeDecl *, 4> knownAssocTypes;
-
-  public:
-    Walker(ProtocolDecl *Proto,
-           llvm::TinyPtrVector<AssociatedTypeDecl *> &assocTypes)
-      : Proto(Proto), assocTypes(assocTypes) {}
-
-    Action walkToTypePre(Type type) override {
-      if (type->is<DependentMemberType>()) {
-        if (auto assocType = getReferencedAssocTypeOfProtocol(type, Proto)) {
-          if (knownAssocTypes.insert(assocType).second)
-            assocTypes.push_back(assocType);
-        }
-
-        return Action::SkipChildren;
-      }
-
-      return Action::Continue;
-    }
-  };
-
-  Walker walker(cast<ProtocolDecl>(req->getDeclContext()), assocTypes);
-
-  // This dance below is to avoid calling getCanonicalType() on a
-  // GenericFunctionType, which creates a GenericSignatureBuilder, which
-  // can in turn trigger associated type inference and cause a cycle.
-  auto reqTy = req->getInterfaceType();
-  if (auto *funcTy = reqTy->getAs<GenericFunctionType>()) {
-    for (auto param : funcTy->getParams())
-      param.getPlainType()->getCanonicalType().walk(walker);
-    funcTy->getResult()->getCanonicalType().walk(walker);
-  } else {
-    reqTy->getCanonicalType().walk(walker);
-  }
-
-  return assocTypes;
-}
-
 void ConformanceChecker::recordWitness(ValueDecl *requirement,
                                        const RequirementMatch &match) {
   // If we already recorded this witness, don't do so again.
@@ -3383,141 +3310,6 @@ static void diagnoseWitnessFixAccessLevel(DiagnosticEngine &diags,
   }
 }
 
-void ConformanceChecker::recordTypeWitness(AssociatedTypeDecl *assocType,
-                                           Type type,
-                                           TypeDecl *typeDecl) {
-
-  // If we already recoded this type witness, there's nothing to do.
-  if (Conformance->hasTypeWitness(assocType)) {
-    assert(Conformance->getTypeWitnessUncached(assocType)
-               .getWitnessType()
-               ->isEqual(type) &&
-           "Conflicting type witness deductions");
-    return;
-  }
-
-  assert(!type->hasArchetype() && "Got a contextual type here?");
-
-  // If there was no type declaration, synthesize one.
-  if (typeDecl == nullptr) {
-    auto aliasDecl = new (getASTContext()) TypeAliasDecl(
-        SourceLoc(), SourceLoc(), assocType->getName(), SourceLoc(),
-        /*genericparams*/ nullptr, DC);
-    aliasDecl->setUnderlyingType(type);
-    
-    aliasDecl->setImplicit();
-    aliasDecl->setSynthesized();
-
-    // Inject the typealias into the nominal decl that conforms to the protocol.
-    auto nominal = DC->getSelfNominalTypeDecl();
-    AccessScope requiredAccessScope = getRequiredAccessScope();
-
-    if (!getASTContext().isSwiftVersionAtLeast(5) &&
-        !DC->getParentModule()->isResilient()) {
-      // HACK: In pre-Swift-5, these typealiases were synthesized with the
-      // same access level as the conforming type, which might be more
-      // visible than the associated type witness. Preserve that behavior
-      // when the underlying type has sufficient access, but only in
-      // non-resilient modules.
-      llvm::Optional<AccessScope> underlyingTypeScope =
-          TypeAccessScopeChecker::getAccessScope(type, DC,
-                                                 /*usableFromInline*/ false);
-      assert(underlyingTypeScope.has_value() &&
-             "the type is already invalid and we shouldn't have gotten here");
-
-      AccessScope nominalAccessScope = nominal->getFormalAccessScope(DC);
-      llvm::Optional<AccessScope> widestPossibleScope =
-          underlyingTypeScope->intersectWith(nominalAccessScope);
-      assert(widestPossibleScope.has_value() &&
-             "we found the nominal and the type witness, didn't we?");
-      requiredAccessScope = widestPossibleScope.value();
-    }
-
-    // An associated type witness can never be less than fileprivate, since
-    // it must always be at least as visible as the enclosing type.
-    AccessLevel requiredAccess =
-        std::max(requiredAccessScope.accessLevelForDiagnostics(),
-                 AccessLevel::FilePrivate);
-
-    aliasDecl->setAccess(requiredAccess);
-    if (isUsableFromInlineRequired()) {
-      auto *attr =
-          new (getASTContext()) UsableFromInlineAttr(/*implicit=*/true);
-      aliasDecl->getAttrs().add(attr);
-    }
-
-    if (nominal == DC) {
-      nominal->addMember(aliasDecl);
-    } else {
-      auto ext = cast<ExtensionDecl>(DC);
-      ext->addMember(aliasDecl);
-    }
-
-    typeDecl = aliasDecl;
-  }
-
-  // Record the type witness.
-  Conformance->setTypeWitness(assocType, type, typeDecl);
-
-  // Record type witnesses for any "overridden" associated types.
-  llvm::SetVector<AssociatedTypeDecl *> overriddenAssocTypes;
-  auto assocOverriddenDecls = assocType->getOverriddenDecls();
-  overriddenAssocTypes.insert(assocOverriddenDecls.begin(),
-                              assocOverriddenDecls.end());
-  for (unsigned idx = 0; idx < overriddenAssocTypes.size(); ++idx) {
-    auto overridden = overriddenAssocTypes[idx];
-
-    // Note all of the newly-discovered overridden associated types.
-    auto overriddenDecls = overridden->getOverriddenDecls();
-    overriddenAssocTypes.insert(overriddenDecls.begin(), overriddenDecls.end());
-
-    // Find the conformance for this overridden protocol.
-    auto overriddenConformance =
-      DC->getParentModule()->lookupConformance(Adoptee,
-                                               overridden->getProtocol(),
-                                               /*allowMissing=*/true);
-    if (overriddenConformance.isInvalid() ||
-        !overriddenConformance.isConcrete())
-      continue;
-
-    auto *overriddenRootConformance =
-        overriddenConformance.getConcrete()->getRootNormalConformance();
-    auto *overriddenRootConformanceDC =
-        overriddenRootConformance->getDeclContext();
-
-    // Don't record a type witness for an overridden associated type if the
-    // conformance to the corresponding inherited protocol
-    // - originates in a superclass
-    // - originates in a different module
-    // - and the current conformance have mismatching conditional requirements
-    // This can turn out badly in two ways:
-    // - Foremost, we must not *alter* conformances originating in superclasses
-    //   or other modules. In other cases, we may hit an assertion in an attempt
-    //   to overwrite an already recorded type witness with a different one.
-    //   For example, the recorded type witness may be invalid, whereas the
-    //   other one---valid, and vice versa.
-    // - If the current conformance is more restrictive, this type witness may
-    //   not be a viable candidate for the overridden associated type.
-    if (overriddenRootConformanceDC->getSelfNominalTypeDecl() !=
-        DC->getSelfNominalTypeDecl())
-      continue;
-
-    if (overriddenRootConformanceDC->getParentModule() != DC->getParentModule())
-      continue;
-
-    auto currConformanceSig = DC->getGenericSignatureOfContext();
-    auto overriddenConformanceSig =
-        overriddenRootConformanceDC->getGenericSignatureOfContext();
-    if (currConformanceSig.getCanonicalSignature() !=
-        overriddenConformanceSig.getCanonicalSignature())
-      continue;
-
-    ConformanceChecker(getASTContext(), overriddenRootConformance,
-                       GlobalMissingWitnesses)
-        .recordTypeWitness(overridden, type, typeDecl);
-  }
-}
-
 /// Whether this protocol is the Objective-C "NSObject" protocol.
 static bool isNSObjectProtocol(ProtocolDecl *proto) {
   if (proto->getNameStr() != "NSObjectProtocol")
@@ -3526,7 +3318,7 @@ static bool isNSObjectProtocol(ProtocolDecl *proto) {
   return proto->hasClangNode();
 }
 
-static Type getTupleConformanceTypeWitness(DeclContext *dc,
+Type swift::getTupleConformanceTypeWitness(DeclContext *dc,
                                            AssociatedTypeDecl *assocType) {
   auto genericSig = dc->getGenericSignatureOfContext();
   assert(genericSig.getGenericParams().size() == 1);
@@ -4625,93 +4417,6 @@ ResolveWitnessResult ConformanceChecker::resolveWitnessViaDefault(
   return ResolveWitnessResult::ExplicitFailed;
 }
 
-# pragma mark Type witness resolution
-
-CheckTypeWitnessResult
-swift::checkTypeWitness(Type type, AssociatedTypeDecl *assocType,
-                        const NormalProtocolConformance *Conf,
-                        SubstOptions options) {
-  auto &ctx = assocType->getASTContext();
-
-  if (type->hasError())
-    return CheckTypeWitnessResult::forError();
-
-  const auto proto = Conf->getProtocol();
-  const auto dc = Conf->getDeclContext();
-  const auto sig = proto->getGenericSignature();
-
-  // FIXME: The RequirementMachine will assert on re-entrant construction.
-  // We should find a more principled way of breaking this cycle.
-  if (ctx.isRecursivelyConstructingRequirementMachine(sig.getCanonicalSignature()) ||
-      ctx.isRecursivelyConstructingRequirementMachine(proto) ||
-      proto->isComputingRequirementSignature())
-    return CheckTypeWitnessResult::forError();
-
-  if (!ctx.LangOpts.hasFeature(Feature::NoncopyableGenerics)
-      && type->isNoncopyable(dc)) {
-    // No move-only type can witness an associatedtype requirement.
-    // Pretend the failure is a lack of Copyable conformance.
-    auto *copyable = ctx.getProtocol(KnownProtocolKind::Copyable);
-    assert(copyable && "missing Copyable protocol!");
-    return CheckTypeWitnessResult::forConformance(copyable);
-  }
-
-  const auto depTy = DependentMemberType::get(proto->getSelfInterfaceType(),
-                                              assocType);
-
-  Type contextType = type->hasTypeParameter() ? dc->mapTypeIntoContext(type)
-                                              : type;
-
-  if (auto superclass = sig->getSuperclassBound(depTy)) {
-    if (superclass->hasTypeParameter()) {
-      // Replace type parameters with other known or tentative type witnesses.
-      superclass = superclass.subst(
-          [&](SubstitutableType *type) {
-            if (type->isEqual(proto->getSelfInterfaceType()))
-              return Conf->getType();
-
-            return Type();
-          },
-          LookUpConformanceInModule(dc->getParentModule()), options);
-
-      if (superclass->hasTypeParameter())
-        superclass = dc->mapTypeIntoContext(superclass);
-    }
-    if (!superclass->isExactSuperclassOf(contextType))
-      return CheckTypeWitnessResult::forSuperclass(superclass);
-  }
-
-  auto *module = dc->getParentModule();
-
-  // Check protocol conformances.
-  for (const auto reqProto : sig->getRequiredProtocols(depTy)) {
-    if (module->lookupConformance(
-            contextType, reqProto,
-            /*allowMissing=*/reqProto->isSpecificProtocol(
-                KnownProtocolKind::Sendable))
-            .isInvalid())
-      return CheckTypeWitnessResult::forConformance(reqProto);
-  }
-
-  if (sig->requiresClass(depTy) &&
-      !contextType->satisfiesClassConstraint()) {
-    return CheckTypeWitnessResult::forLayout(
-        module->getASTContext().getAnyObjectType());
-  }
-
-  // Tuple conformances can only witness associated types by projecting them
-  // element-wise.
-  if (isa<BuiltinTupleDecl>(dc->getSelfNominalTypeDecl())) {
-    auto expectedTy = getTupleConformanceTypeWitness(dc, assocType);
-    if (!expectedTy->isEqual(type)) {
-      return CheckTypeWitnessResult::forTuple(expectedTy);
-    }
-  }
-
-  // Success!
-  return CheckTypeWitnessResult::forSuccess();
-}
-
 ResolveWitnessResult
 ConformanceChecker::resolveWitnessTryingAllStrategies(ValueDecl *requirement) {
   decltype(&ConformanceChecker::resolveWitnessViaLookup) strategies[] = {
@@ -4734,205 +4439,53 @@ ConformanceChecker::resolveWitnessTryingAllStrategies(ValueDecl *requirement) {
   return ResolveWitnessResult::Missing;
 }
 
-/// Attempt to resolve a type witness via member name lookup.
-ResolveWitnessResult ConformanceChecker::resolveTypeWitnessViaLookup(
-                       AssociatedTypeDecl *assocType) {
-  // Conformances constructed by the ClangImporter should have explicit type
-  // witnesses already.
-  if (isa<ClangModuleUnit>(Conformance->getDeclContext()->getModuleScopeContext())) {
-    llvm::errs() << "Cannot look up associated type for imported conformance:\n";
-    Conformance->getType().dump(llvm::errs());
-    assocType->dump(llvm::errs());
-    abort();
+void ConformanceChecker::resolveSingleWitness(ValueDecl *requirement) {
+  assert(!isa<AssociatedTypeDecl>(requirement) && "Not a value witness");
+  assert(!Conformance->hasWitness(requirement) && "Already resolved");
+
+  // Note that we're resolving this witness.
+  assert(ResolvingWitnesses.count(requirement) == 0 && "Currently resolving");
+  ResolvingWitnesses.insert(requirement);
+  SWIFT_DEFER { ResolvingWitnesses.erase(requirement); };
+
+  // Make sure we've validated the requirement.
+  if (requirement->isInvalid()) {
+    Conformance->setInvalid();
+    return;
   }
 
-  NLOptions subOptions = (NL_QualifiedDefault | NL_OnlyTypes | NL_ProtocolMembers);
+  if (!requirement->isProtocolRequirement())
+    return;
 
-  // Look for a member type with the same name as the associated type.
-  SmallVector<ValueDecl *, 4> candidates;
+  // Resolve all associated types before trying to resolve this witness.
+  resolveTypeWitnesses();
 
-  DC->lookupQualified(DC->getSelfNominalTypeDecl(),
-                      assocType->createNameRef(),
-                      DC->getSelfNominalTypeDecl()->getLoc(),
-                      subOptions, candidates);
+  // If any of the type witnesses was erroneous, don't bother to check
+  // this value witness: it will fail.
+  auto assocTypes = evaluateOrDefault(getASTContext().evaluator,
+                                      ReferencedAssociatedTypesRequest{requirement},
+                                      TinyPtrVector<AssociatedTypeDecl *>());
 
-  // If there aren't any candidates, we're done.
-  if (candidates.empty()) {
-    return ResolveWitnessResult::Missing;
-  }
-
-  // Determine which of the candidates is viable.
-  SmallVector<LookupTypeResultEntry, 2> viable;
-  SmallVector<std::pair<TypeDecl *, CheckTypeWitnessResult>, 2> nonViable;
-  SmallPtrSet<CanType, 4> viableTypes;
-
-  for (auto candidate : candidates) {
-    auto *typeDecl = cast<TypeDecl>(candidate);
-
-    // Skip other associated types.
-    if (isa<AssociatedTypeDecl>(typeDecl))
-      continue;
-
-    auto *genericDecl = cast<GenericTypeDecl>(typeDecl);
-
-    // If the declaration has generic parameters, it cannot witness an
-    // associated type.
-    if (genericDecl->isGeneric())
-      continue;
-
-    // As a narrow fix for a source compatibility issue with SwiftUI's
-    // swiftinterface, allow a 'typealias' type witness with an underlying type
-    // of 'Never' if it is declared in a context that does not satisfy the
-    // requirements of the conformance context.
-    //
-    // FIXME: Remove this eventually.
-    bool skipRequirementCheck = false;
-    if (auto *typeAliasDecl = dyn_cast<TypeAliasDecl>(typeDecl)) {
-      if (typeAliasDecl->getParentModule()->getName().is("SwiftUI") &&
-          typeAliasDecl->getParentSourceFile() &&
-          typeAliasDecl->getParentSourceFile()->Kind == SourceFileKind::Interface) {
-        if (typeAliasDecl->getUnderlyingType()->isNever()) {
-          if (typeAliasDecl->getDeclContext()->getSelfNominalTypeDecl() ==
-              DC->getSelfNominalTypeDecl()) {
-            skipRequirementCheck = true;
-          }
-        }
-      }
-    }
-
-    // Skip typealiases with an unbound generic type as their underlying type.
-    if (auto *typeAliasDecl = dyn_cast<TypeAliasDecl>(typeDecl))
-      if (typeAliasDecl->getDeclaredInterfaceType()->is<UnboundGenericType>())
-        continue;
-
-    // Skip dependent protocol typealiases.
-    //
-    // FIXME: This should not be necessary.
-    if (auto *typeAliasDecl = dyn_cast<TypeAliasDecl>(typeDecl)) {
-      if (isa<ProtocolDecl>(typeAliasDecl->getDeclContext()) &&
-          typeAliasDecl->getUnderlyingType()->getCanonicalType()
-            ->hasTypeParameter()) {
-        continue;
-      }
-    }
-
-    // If the type comes from a constrained extension or has a 'where'
-    // clause, check those requirements now.
-    if (!skipRequirementCheck &&
-        !TypeChecker::checkContextualRequirements(
-            genericDecl, Adoptee, SourceLoc(), DC->getParentModule(),
-            DC->getGenericSignatureOfContext())) {
-      continue;
-    }
-
-    auto memberType = TypeChecker::substMemberTypeWithBase(DC->getParentModule(),
-                                                           typeDecl, Adoptee);
-
-    // Type witnesses that resolve to constraint types are always
-    // existential types. This can only happen when the type witness
-    // is explicitly written with a type alias. The type alias itself
-    // is still a constraint type because it can be used as both a
-    // type witness and as a generic constraint.
-    //
-    // With SE-0335, using a type alias as both a type witness and a generic
-    // constraint will be disallowed in Swift 6, because existential types
-    // must be explicit, and a generic constraint isn't a valid type witness.
-    if (memberType->isConstraintType()) {
-      memberType = ExistentialType::get(memberType);
-    }
-
-    if (!viableTypes.insert(memberType->getCanonicalType()).second)
-      continue;
-
-    // Check this type against the protocol requirements.
-    if (auto checkResult =
-            checkTypeWitness(memberType, assocType, Conformance)) {
-      nonViable.push_back({typeDecl, checkResult});
-    } else {
-      viable.push_back({typeDecl, memberType, nullptr});
+  for (auto assocType : assocTypes) {
+    if (Conformance->getTypeWitness(assocType)->hasError()) {
+      Conformance->setInvalid();
+      return;
     }
   }
 
-  // If there are no viable witnesses, and all nonviable candidates came from
-  // protocol extensions, treat this as "missing".
-  if (viable.empty() &&
-      std::find_if(nonViable.begin(), nonViable.end(),
-                   [](const std::pair<TypeDecl *, CheckTypeWitnessResult> &x) {
-                     return x.first->getDeclContext()
-                        ->getSelfProtocolDecl() == nullptr;
-                   }) == nonViable.end())
-    return ResolveWitnessResult::Missing;
+  // Try to resolve the witness.
+  switch (resolveWitnessTryingAllStrategies(requirement)) {
+  case ResolveWitnessResult::Success:
+    return;
 
-  // If there is a single viable candidate, form a substitution for it.
-  if (viable.size() == 1) {
-    auto interfaceType = viable.front().MemberType;
-    recordTypeWitness(assocType, interfaceType, viable.front().Member);
-    return ResolveWitnessResult::Success;
+  case ResolveWitnessResult::ExplicitFailed:
+    Conformance->setInvalid();
+    recordInvalidWitness(requirement);
+    return;
+
+  case ResolveWitnessResult::Missing:
+    llvm_unreachable("Should have failed");
   }
-
-  // Record an error.
-  recordTypeWitness(assocType, ErrorType::get(getASTContext()), nullptr);
-
-  // If we had multiple viable types, diagnose the ambiguity.
-  if (!viable.empty()) {
-    getASTContext().addDelayedConformanceDiag(Conformance, true,
-      [assocType, viable](NormalProtocolConformance *conformance) {
-        auto &diags = assocType->getASTContext().Diags;
-        diags.diagnose(assocType, diag::ambiguous_witnesses_type,
-                       assocType->getName());
-
-        for (auto candidate : viable)
-          diags.diagnose(candidate.Member, diag::protocol_witness_type);
-      });
-
-    return ResolveWitnessResult::ExplicitFailed;
-  }
-  // Save the missing type witness for later diagnosis.
-  GlobalMissingWitnesses.insert({assocType, {}});
-
-  // None of the candidates were viable.
-  getASTContext().addDelayedConformanceDiag(Conformance, true,
-    [nonViable](NormalProtocolConformance *conformance) {
-      auto &diags = conformance->getDeclContext()->getASTContext().Diags;
-      for (auto candidate : nonViable) {
-        if (candidate.first->getDeclaredInterfaceType()->hasError() ||
-            candidate.second.getKind() == CheckTypeWitnessResult::Error)
-          continue;
-
-        switch (candidate.second.getKind()) {
-        case CheckTypeWitnessResult::Success:
-        case CheckTypeWitnessResult::Error:
-          llvm_unreachable("Should not end up here");
-
-        case CheckTypeWitnessResult::Conformance:
-        case CheckTypeWitnessResult::Layout:
-          diags.diagnose(
-             candidate.first,
-             diag::protocol_type_witness_unsatisfied_conformance,
-             candidate.first->getDeclaredInterfaceType(),
-             candidate.second.getRequirement());
-          break;
-
-        case CheckTypeWitnessResult::Superclass:
-          diags.diagnose(
-             candidate.first,
-             diag::protocol_type_witness_unsatisfied_superclass,
-             candidate.first->getDeclaredInterfaceType(),
-             candidate.second.getRequirement());
-          break;
-
-        case CheckTypeWitnessResult::Tuple:
-          diags.diagnose(
-             candidate.first,
-             diag::protocol_type_witness_tuple,
-             candidate.first->getDeclaredInterfaceType(),
-             candidate.second.getRequirement());
-          break;
-        }
-      }
-    });
-
-  return ResolveWitnessResult::ExplicitFailed;
 }
 
 /// FIXME: It feels like this could be part of findExistentialSelfReferences().
@@ -6906,25 +6459,6 @@ swift::findWitnessedObjCRequirements(const ValueDecl *witness,
   return result;
 }
 
-TypeWitnessAndDecl
-TypeWitnessRequest::evaluate(Evaluator &eval,
-                             NormalProtocolConformance *conformance,
-                             AssociatedTypeDecl *requirement) const {
-  llvm::SetVector<ASTContext::MissingWitness> MissingWitnesses;
-  ConformanceChecker checker(requirement->getASTContext(), conformance,
-                             MissingWitnesses);
-  checker.resolveSingleTypeWitness(requirement);
-  checker.diagnoseMissingWitnesses(MissingWitnessDiagnosisKind::ErrorFixIt,
-                                   /*Delayed=*/true);
-  // FIXME: ConformanceChecker and the other associated WitnessCheckers have
-  // an extremely convoluted caching scheme that doesn't fit nicely into the
-  // evaluator's model. All of this should be refactored away.
-  const auto known = conformance->TypeWitnesses.find(requirement);
-  assert(known != conformance->TypeWitnesses.end() &&
-         "Didn't resolve witness?");
-  return known->second;
-}
-
 Witness
 ValueWitnessRequest::evaluate(Evaluator &eval,
                               NormalProtocolConformance *conformance,
@@ -6945,42 +6479,6 @@ ValueWitnessRequest::evaluate(Evaluator &eval,
     return Witness();
   }
   return known->second;
-}
-
-ProtocolConformanceRef
-AssociatedConformanceRequest::evaluate(Evaluator &eval,
-                                       NormalProtocolConformance *conformance,
-                                       CanType origTy, ProtocolDecl *reqProto,
-                                       unsigned index) const {
-  auto *module = conformance->getDeclContext()->getParentModule();
-
-  auto subMap = SubstitutionMap::getProtocolSubstitutions(
-      conformance->getProtocol(),
-      conformance->getType(),
-      ProtocolConformanceRef(conformance));
-  auto substTy = origTy.subst(subMap);
-
-  // Looking up a conformance for a contextual type and mapping the
-  // conformance context produces a more accurate result than looking
-  // up a conformance from an interface type.
-  //
-  // This can happen if the conformance has an associated conformance
-  // depending on an associated type that is made concrete in a
-  // refining protocol.
-  //
-  // That is, the conformance of an interface type G<T> : P really
-  // depends on the generic signature of the current context, because
-  // performing the lookup in a "more" constrained extension than the
-  // one where the conformance was defined must produce concrete
-  // conformances.
-  //
-  // FIXME: Eliminate this, perhaps by adding a variant of
-  // lookupConformance() taking a generic signature.
-  if (substTy->hasTypeParameter())
-    substTy = conformance->getDeclContext()->mapTypeIntoContext(substTy);
-
-  return module->lookupConformance(substTy, reqProto, /*allowMissing=*/true)
-      .mapConformanceOutOfContext();
 }
 
 ValueDecl *TypeChecker::deriveProtocolRequirement(DeclContext *DC,
@@ -7060,41 +6558,6 @@ ValueDecl *TypeChecker::deriveProtocolRequirement(DeclContext *DC,
   llvm_unreachable("unknown derivable protocol kind");
 }
 
-std::pair<Type, TypeDecl *>
-TypeChecker::deriveTypeWitness(DeclContext *DC,
-                               NominalTypeDecl *TypeDecl,
-                               AssociatedTypeDecl *AssocType) {
-  auto *protocol = cast<ProtocolDecl>(AssocType->getDeclContext());
-
-  auto knownKind = protocol->getKnownProtocolKind();
-  
-  if (!knownKind)
-    return std::make_pair(nullptr, nullptr);
-
-  auto Decl = DC->getInnermostDeclarationDeclContext();
-
-  DerivedConformance derived(TypeDecl->getASTContext(), Decl, TypeDecl,
-                             protocol);
-  switch (*knownKind) {
-  case KnownProtocolKind::RawRepresentable:
-    return std::make_pair(derived.deriveRawRepresentable(AssocType), nullptr);
-  case KnownProtocolKind::CaseIterable:
-    return std::make_pair(derived.deriveCaseIterable(AssocType), nullptr);
-  case KnownProtocolKind::Differentiable:
-    return derived.deriveDifferentiable(AssocType);
-  case KnownProtocolKind::DistributedActor:
-    return derived.deriveDistributedActor(AssocType);
-  case KnownProtocolKind::Identifiable:
-    // Identifiable only has derivation logic for distributed actors,
-    // because how it depends on the ActorSystem the actor is associated with.
-    // If the nominal wasn't a distributed actor, we should not end up here,
-    // but either way, then we'd return null (fail derivation).
-    return derived.deriveDistributedActor(AssocType);
-  default:
-    return std::make_pair(nullptr, nullptr);
-  }
-}
-
 namespace {
   class DefaultWitnessChecker : public WitnessChecker {
     
@@ -7152,9 +6615,7 @@ void TypeChecker::inferDefaultWitnesses(ProtocolDecl *proto) {
   // Find the default for the given associated type.
   auto findAssociatedTypeDefault = [proto](AssociatedTypeDecl *assocType)
       -> std::pair<Type, AssociatedTypeDecl *> {
-    auto defaultedAssocType =
-        AssociatedTypeInference::findDefaultedAssociatedType(
-            proto, proto, assocType);
+    auto defaultedAssocType = findDefaultedAssociatedType(proto, proto, assocType);
     if (!defaultedAssocType)
       return {Type(), nullptr};
